feat(clients): add rTorrent support
