+++ go.mod
@@ -20,7 +20,9 @@ require (
 	github.com/hekmon/transmissionrpc/v2 v2.0.1
 	github.com/lib/pq v1.10.6
 	github.com/mattn/go-shellwords v1.0.12
+	github.com/mmcdole/gofeed v1.1.3
 	github.com/moistari/rls v0.2.6
+	github.com/mrobinsn/go-rtorrent v1.8.0
 	github.com/pkg/errors v0.9.1
 	github.com/r3labs/sse/v2 v2.8.1
 	github.com/robfig/cron/v3 v3.0.1
@@ -63,7 +65,6 @@ require (
 	github.com/mattn/go-colorable v0.1.12 // indirect
 	github.com/mattn/go-isatty v0.0.14 // indirect
 	github.com/mitchellh/mapstructure v1.5.0 // indirect
-	github.com/mmcdole/gofeed v1.1.3 // indirect
 	github.com/mmcdole/goxpp v0.0.0-20181012175147-0068e33feabf // indirect
 	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
 	github.com/modern-go/reflect2 v1.0.2 // indirect
