+++ web/src/domain/constants.ts
@@ -177,6 +177,11 @@ export const DownloadClientTypeOptions: RadioFieldsetOption[] = [
     value: "DELUGE_V2"
   },
   {
+    label: "rTorrent",
+    description: "Add torrents directly to rTorrent",
+    value: "RTORRENT"
+  },
+  {
     label: "Transmission",
     description: "Add torrents directly to Transmission",
     value: "TRANSMISSION"
@@ -207,6 +212,7 @@ export const DownloadClientTypeNameMap: Record<DownloadClientType | string, stri
   "DELUGE_V1": "Deluge v1",
   "DELUGE_V2": "Deluge v2",
   "QBITTORRENT": "qBittorrent",
+  "RTORRENT": "rTorrent",
   "TRANSMISSION": "Transmission",
   "RADARR": "Radarr",
   "SONARR": "Sonarr",
@@ -222,6 +228,7 @@ export const ActionTypeOptions: RadioFieldsetOption[] = [
   { label: "qBittorrent", description: "Add torrents directly to qBittorrent", value: "QBITTORRENT" },
   { label: "Deluge", description: "Add torrents directly to Deluge", value: "DELUGE_V1" },
   { label: "Deluge v2", description: "Add torrents directly to Deluge 2", value: "DELUGE_V2" },
+  { label: "rTorrent", description: "Add torrents directly to rTorrent", value: "RTORRENT" },
   { label: "Transmission", description: "Add torrents directly to Transmission", value: "TRANSMISSION" },
   { label: "Radarr", description: "Send to Radarr and let it decide", value: "RADARR" },
   { label: "Sonarr", description: "Send to Sonarr and let it decide", value: "SONARR" },
@@ -237,6 +244,7 @@ export const ActionTypeNameMap = {
   "DELUGE_V1": "Deluge v1",
   "DELUGE_V2": "Deluge v2",
   "QBITTORRENT": "qBittorrent",
+  "RTORRENT": "rTorrent",
   "TRANSMISSION": "Transmission",
   "RADARR": "Radarr",
   "SONARR": "Sonarr",
@@ -247,7 +255,7 @@ export const ActionTypeNameMap = {
 export const ActionContentLayoutOptions: SelectGenericOption<ActionContentLayout>[] = [
   { label: "Original", description: "Original", value: "ORIGINAL" },
   { label: "Create subfolder", description: "Create subfolder", value: "SUBFOLDER_CREATE" },
-  { label: "Don't create subfolder", description: "Don't create subfolder", value: "SUBFOLDER_NONE" },
+  { label: "Don't create subfolder", description: "Don't create subfolder", value: "SUBFOLDER_NONE" }
 ];
 
 export interface OptionBasic {
